// backend/apps/solicitud/src/domain/models/salescreate.model.ts
/* eslint-disable prettier/prettier */
export class IngresaVentaModel {
  Id_Solicitud?: string | null;
  ID_PRODUCTO?: string | null;
  PLAZO?: number;
  CUOTA?: number;
  CICLO?: number;
  PK_PSE_CAT_PROMOCION_PLAZO?: number;
  FechaLocal?: Date;
  Usuario?: string | null;
  LIMITE_CREDITO?: number;
  MONTO_VENTA?: number;
  TASA_INTERES_NORMAL?: number;
}